Add aspect ratio checking
The frame has a 4:3 aspect ratio (1.5), so we'll allow pictures that have a 1.4
to 1.6 aspect ratio, and otherwise return a notification to the user that the
uploaded photo has an aspect ratio that would not work well in the frame. This
validation will work regardless of portrait or landscape orientation.

This should also catch a class of UI weirdness where images are too tall and
push content below them too far down. This does NOT adjust the frame for
portrait photos though; everything is still treated as landscape.